Remove `makeStyles` from `AlkemioAvatar` and `CommunityUpdatesView` (#7390)
* Clean up alkemio avatar component

The goal is to remove `@mui/styles`, but as the component is deprecated,
it is replaced with the new avatar component in the user popup and the
deprecated component is cleaned up from all unused features. The user
popup is inlined in it to simplify more the interface of the
component.