Remove `@mui/styles` legacy package (#7446)
* Remove makeStyles for global styles

MUI v5 already has a component designed for holding global styles, that
is used instead of the custom one.

* Remove makeStyles from QR code dialogs

* Remove makeStyles from deprecated section

* Remove makeStyles from loading

* Remove makeStyles from CredentialProfileCard

* Delete deprecated tags component

It was only used in the UserPopUp modal. The styles are a little
different with the TagsComponent, but they look better.

* Clean up UserPopup dialog

* makesStyles are replaced with styled components
* deprecated components are replaced

The styles are slightly changed, but the whole dialog seems for
refactoring.

* Credential card - use styled components instead of makeStyles

Clean up some more User popup - use MUI divider instead of custom.

* Auth required page - remove makeStyles

* Clean up deprecated section component

* Delete unused properties
* Remove makeStyles

* WrapperBackdrop - use styled component instead of makeStyles

* Fix problem with styled_default is not a function

The problem came from calling the MUI Autocomplete component the React
Vite app.

* Remove makeStyles for Whiteboard dialogs

Also fix the paper dialog styles for QR code dialog and request
credentials dialog, as they were not applied to the correct element.

* Remove makeStyles from Application dialog

* Delete useProfileStyles

It is used only in MarkdownInput, but the class name seems not applied.

* Refactor membership backdrop

Inline Backdrop with message component and remove makeStyles.

* Remove makeStyles from contributor components

* Delete SearchableListIconButton

* Clean up circle tag

Clean up the component from unused props. Delete unused component -
MetricCircleView. Replace makeStyles with sx and WrapperTypology with
Typology.

* Remove makeStyles for FormikSelect

* Fix wrong import of styled

The styled import was from the deprecated @mui/styles package. It is
updated to use styled from @mui/material/styles.

* Remove makeStyles from Excalidraw whiteboard

* Replace deprecated WrapperTypography in some pages

Also rework Links list not to use @mui/styles.

* Delete WrapperTypography - replace with Typography

* Delete legacy package '@mui/styles'

* Add alt text to section image banner

* Import DialogContent from @mui/material directly

Delete the `src/core/ui/dialog/DialogContent.ts` file and update all
related imports.

* Remove unused prop - message

* Add types to MetricViewItem

`count` is never undefined, so it is changed to mandatory property.

* Address some coderabbitai comments

* excalidraw help dialog z-index fix

* Extract the repeating styles in object

* Delete file after merge conflicts

* fixup! Extract the repeating styles in object

---------

Co-authored-by: Bobby Kolev <bobbykolev@abv.bg>