Refactors space list and item components (#8058)
* Refactors space list and item components

Updates the space list to improve performance and readability.

Moves authorization check to SpaceListItem to simplify SpaceList.
Disables form elements when saving or lacking update permissions.

Also, increments package version.

* Prevents saving space settings without update rights

Disables the save button in the space settings modal if the user does not have update permissions for the space. This prevents users without proper authorization from attempting to save changes that they are not allowed to make.

* Uses `includes` for privilege check.

Updates the privilege check to use `includes` for better readability and efficiency.

This change replaces the `find` method with `includes` to determine if a specific authorization privilege exists within the user's privileges. `includes` provides a more concise and direct way to check for the presence of an element in an array, improving code clarity.