Override for link recovery flow (#8700)
* Enhances password recovery message

Improves clarity of password recovery messages by dynamically adjusting the displayed text based on whether a link or code is sent.

This change ensures users understand what type of recovery method to expect, improving the overall user experience during password recovery.

* Makes recovery type translatable

Replaces the hardcoded "code" string with a translation key
in password recovery messages for multiple languages.

This allows the recovery type to be dynamically updated,
for example, to display "link" instead of "code" in the
message.

* Fixes typo in password recovery message

Corrects a typo in the "initial" message under the "recovery" section in the English translation file.
The phrase "allow to reset" is changed to "allow you to reset" for better grammar.
The other translation files are updated with the same text.

* Translates password recovery message

Updates the password recovery message in Spanish and French translations for better clarity and user experience.

The updated message provides clearer instructions for users who have requested a password recovery, including checking spam folders and alternative login methods.

* Updates recovery message for clarity

Improves the wording of the password recovery message
to use "link" instead of a generic "type", enhancing
user understanding.